refactor: group specialized evaluators under rllm/eval/evaluator/
Move the 6 standalone evaluator implementations into a dedicated
sub-package and drop the redundant `_evaluator` suffix:

  rllm/eval/bfcl_evaluator.py        → rllm/eval/evaluator/bfcl.py
  rllm/eval/ifeval_evaluator.py      → rllm/eval/evaluator/ifeval.py
  rllm/eval/llm_equality_evaluator.py → rllm/eval/evaluator/llm_equality.py
  rllm/eval/llm_judge_evaluator.py   → rllm/eval/evaluator/llm_judge.py
  rllm/eval/translation_evaluator.py → rllm/eval/evaluator/translation.py
  rllm/eval/widesearch_evaluator.py  → rllm/eval/evaluator/widesearch.py

Add rllm/eval/evaluator/__init__.py that re-exports all six classes for
short-form imports. Update evaluator_loader.py and the affected tests.

Protocol definitions (AgentFlow, Evaluator) and lightweight built-in
evaluators (Math, Countdown, Code, F1, MCQ, IoU, PointInMask, Depth,
Compound) stay in rllm/eval/types.py — they're tied to the protocol
and don't pull heavy deps.
