+++ rllm/eval/evaluator/__init__.py
@@ -1,0 +1,24 @@
+"""Specialized evaluators for specific benchmarks.
+
+These complement the lightweight built-in evaluators (Math, Countdown,
+Code, F1, MCQ, IoU, etc.) defined in :mod:`rllm.eval.types`. The ones
+here pull in heavier dependencies or implement domain-specific scoring
+(BFCL function-call matching, IFEval constraint checking, LLM judges,
+translation chrF, WideSearch grading).
+"""
+
+from rllm.eval.evaluator.bfcl import BFCLEvaluator
+from rllm.eval.evaluator.ifeval import IFEvalEvaluator
+from rllm.eval.evaluator.llm_equality import LLMEqualityEvaluator
+from rllm.eval.evaluator.llm_judge import LLMJudgeEvaluator
+from rllm.eval.evaluator.translation import TranslationEvaluator
+from rllm.eval.evaluator.widesearch import WideSearchEvaluator
+
+__all__ = [
+    "BFCLEvaluator",
+    "IFEvalEvaluator",
+    "LLMEqualityEvaluator",
+    "LLMJudgeEvaluator",
+    "TranslationEvaluator",
+    "WideSearchEvaluator",
+]
